Address some issues reported by Undefined Behavior Sanitizer
The runtime warnings were:
- null pointer passed as argument NNN, which is declared to never be null
- left shift of MMM by NNN places cannot be represented in type 'int'

+++ src/e-util/e-bit-array.c
@@ -224,7 +224,7 @@ e_bit_array_foreach (EBitArray *bit_array,
 	}
 }
 
-#define PART(x,n) (((x) & (0x01010101 << n)) >> n)
+#define PART(x,n) ((guint32) ((((guint64) x) & (((guint64) 0x01010101) << n)) >> n))
 #define SECTION(x, n) (((x) >> (n * 8)) & 0xff)
 
 /**
